+++ configure.ac
@@ -21,6 +21,11 @@ if test -d ".git" ; then
 fi
 AC_MSG_NOTICE([RPM_RELEASE='$RPM_RELEASE'])
 
+AC_CHECK_PROG(SPHINX_BUILD, sphinx-build, yes, no)
+if test x"$SPHINX_BUILD" = x"no"; then
+  AC_MSG_ERROR([sphinx-build not found (python-sphinx)])
+fi
+
 AC_CONFIG_MACRO_DIR([m4])
 
 AC_CONFIG_SUBDIRS([src/gmock])
