+++ debian/control
@@ -46,6 +46,7 @@ Build-Depends: autoconf,
                python (>= 2.6.6-3~),
                python-argparse,
                python-nose,
+               python-sphinx,
                python-virtualenv,
                sdparm | hdparm,
                uuid-dev,
