+++ doc/conf.py
@@ -32,43 +32,6 @@ extensions = [
     ]
 todo_include_todos = True
 
-def _get_manpages():
-    import os
-    man_dir = os.path.join(
-        os.path.dirname(__file__),
-        'man',
-        )
-    sections = os.listdir(man_dir)
-    for section in sections:
-        section_dir = os.path.join(man_dir, section)
-        if not os.path.isdir(section_dir):
-            continue
-        for filename in os.listdir(section_dir):
-            base, ext = os.path.splitext(filename)
-            if ext != '.rst':
-                continue
-            if base == 'index':
-                continue
-            with file(os.path.join(section_dir, filename)) as f:
-                one = f.readline()
-                two = f.readline()
-                three = f.readline()
-                assert one == three
-                assert all(c=='=' for c in one.rstrip('\n'))
-                two = two.strip()
-                name, rest = two.split('--', 1)
-                assert name.strip() == base
-                description = rest.strip()
-            yield (
-                os.path.join('man', section, base),
-                base,
-                description,
-                '',
-                section,
-                )
-
-man_pages = list(_get_manpages())
-
 top_level = os.path.dirname(
     os.path.dirname(
         os.path.abspath(__file__)
