+++ man/8/ceph-rest-api.rst
