+++ man/8/librados-config.rst
