+++ man/8/rados.rst
